#1289 Iterator refactor
* Replace ecs_iter_t::columns with ecs_table_record_t* array

* Remove ecs_iter_t::ptrs

* Update C++ API to new iterator design

* Fix warnings

+++ src/observable.h
@@ -32,7 +32,7 @@ typedef struct ecs_observer_impl_t {
     int32_t last_event_id_storage;
 
     ecs_id_t register_id;       /**< Id observer is registered with (single term observers only) */
-    int32_t term_index;         /**< Index of the term in parent observer (single term observers only) */
+    int8_t term_index;          /**< Index of the term in parent observer (single term observers only) */
 
     ecs_flags32_t flags;        /**< Observer flags */
     uint64_t id;                /**< Internal id (not entity id) */
